Move partly revised extraction engine into SOSS/dms.
